Reformat and delete redundant code. fix functional issue
