+++ cluster.html
@@ -1,30 +1,34 @@
 <!DOCTYPE html>
 <html>
-<head>
-    <meta charset="utf-8"/>
-    <meta name="viewport" content="initial-scale=1,maximum-scale=1,user-scalable=no"/>
+    <head>
+        <meta charset="utf-8" />
+        <meta name="viewport"
+            content="initial-scale=1,maximum-scale=1,user-scalable=no" />
 
-    <title>TLCMap Cluster Viewer</title>
+        <title>TLCMap Cluster Viewer</title>
 
-    <link rel="stylesheet" href="./view.css">
-    <link rel="stylesheet" href="https://js.arcgis.com/4.22/esri/themes/light/main.css"/>
-</head>
+        <link rel="stylesheet" href="./view.css">
+        <link rel="stylesheet"
+            href="https://js.arcgis.com/4.22/esri/themes/light/main.css" />
+    </head>
 
-<body>
-<div id="viewDiv"></div>
+    <body>
+        <div id="viewDiv"></div>
 
-<div id="infoDiv">
-    <a href="https://www.tlcmap.org"><img class="mdicon"></a>
-</div>
+        <div id="infoDiv">
+            <a href="https://www.tlcmap.org"><img class="mdicon"></a>
+        </div>
 
-<script src="https://cdnjs.cloudflare.com/ajax/libs/dompurify/2.3.3/purify.min.js"></script>
-<script src="https://js.arcgis.com/4.22/"></script>
+        <script
+            src="https://cdnjs.cloudflare.com/ajax/libs/dompurify/2.3.3/purify.min.js"></script>
+        <script src="https://js.arcgis.com/4.22/"></script>
 
-<!-- mainly just for getting the layers json metadata. -->
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-<script src="configLoader.js"></script>
-<script src="mapConfigurator.js"></script>
-<script src="cluster.js"></script>
+        <!-- mainly just for getting the layers json metadata. -->
+        <script
+            src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+        <script src="config-loader.js"></script>
+        <script src="map-configurator.js"></script>
+        <script src="cluster.js"></script>
 
-</body>
+    </body>
 </html>
