+++ collection-3d.html
@@ -1,31 +1,34 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="initial-scale=1,maximum-scale=1,user-scalable=no"/>
-    <link rel="stylesheet" href="./view.css"/>
-    <link rel="stylesheet" href="collection-legend.css"/>
-    <link rel="stylesheet" href="https://js.arcgis.com/4.22/esri/themes/light/main.css"/>
-    <title>TLCMap Multilayer 3D Viewer</title>
-</head>
-<body>
-<div id="viewDiv"></div>
-<div id="infoDiv">
-    <a href="https://www.tlcmap.org"><img class="mdicon"></a>
-</div>
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport"
+            content="initial-scale=1,maximum-scale=1,user-scalable=no" />
+        <link rel="stylesheet" href="./view.css" />
+        <link rel="stylesheet" href="collection-legend.css" />
+        <link rel="stylesheet"
+            href="https://js.arcgis.com/4.22/esri/themes/light/main.css" />
+        <title>TLCMap Multilayer 3D Viewer</title>
+    </head>
+    <body>
+        <div id="viewDiv"></div>
+        <div id="infoDiv">
+            <a href="https://www.tlcmap.org"><img class="mdicon"></a>
+        </div>
 
+        <script
+            src="https://cdnjs.cloudflare.com/ajax/libs/dompurify/2.3.3/purify.min.js"></script>
+        <script src="https://js.arcgis.com/4.22/"></script>
+        <script
+            src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
 
-<script src="https://cdnjs.cloudflare.com/ajax/libs/dompurify/2.3.3/purify.min.js"></script>
-<script src="https://js.arcgis.com/4.22/"></script>
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+        <script src="legend-color-generateor.js"></script>
+        <script src="collection-utility.js"></script>
+        <script src="collection-legend.js"></script>
 
-<script src="legend-color-generateor.js"></script>
-<script src="collection-utility.js"></script>
-<script src="collection-legend.js"></script>
-
-<script src="collectionConfigLoader.js"></script>
-<script src="configLoader.js"></script>
-<script src="mapConfigurator.js"></script>
-<script src="collection-3d.js"></script>
-</body>
+        <script src="collection-config-loader.js"></script>
+        <script src="config-loader.js"></script>
+        <script src="map-configurator.js"></script>
+        <script src="collection-3d.js"></script>
+    </body>
 </html>
