+++ view.css
@@ -28,7 +28,6 @@ img.mdicon {
     margin-left: auto;
     margin-right: auto;
     width: 50px;
-    height: 50px;
 }
 
 #timeSlider {
